Split main file into module files
